move read prompt to lab.utils
