+++ src/lab/mcp/__init__.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from ._typing import CommonMessage, ToolInfo, ToolMessage
 from .client import MCPHandlerInterface, TimeemiMCPHandler
-from .util import read_prompt_from_mcp_prompt_template, read_prompt_from_text_file, read_result_from_mcp_tool_response
+from .util import read_prompt_from_mcp_prompt_template, read_result_from_mcp_tool_response
 from .virtual_client import VirtualMCPHandler, get_virtual_mcp_handler, test_virtual_mcp_handler
 
 __all__ = [
@@ -14,7 +14,6 @@ __all__ = [
     "MCPHandlerInterface",
     "TimeemiMCPHandler",
     "read_prompt_from_mcp_prompt_template",
-    "read_prompt_from_text_file",
     "read_result_from_mcp_tool_response",
     "VirtualMCPHandler",
 ]
